requirements.txt
Pillow==5.4.1
click==6.7
